Fast is false unless string is true
